requirements.txt
flask>=3.0.0

cryptography>=42.0.0
flask-limiter>=3.5.0
redis>=5.0.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Phase 2 dependencies
opencv-python>=4.9.0.80
pytesseract>=0.3.10
pillow>=10.2.0
celery>=5.3.4
twilio>=8.13.0
numpy>=1.26.3

# Phase 3 dependencies
sentence-transformers>=2.3.1
google-generativeai>=0.3.2
googlemaps>=4.10.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
requests>=2.31.0
